Refactor CreateWorkLocation and UpdateWorkLocation components to use WorkLocationEmployeeSelector; implement unsaved changes guard; enhance validation schema for work location; update API query keys and hooks for employee data retrieval.
