+++ frontend/src/community/configurations/components/organisms/UpdateWorkLocation/UpdateWorkLocation.tsx
@@ -1,41 +1,23 @@
 import { useFormik } from "formik";
-import * as Yup from "yup";
 import { ButtonV2, InputField } from "@rootcodelabs/skapp-ui";
-import { Box, Checkbox as MuiCheckbox, Typography } from "@mui/material";
 import { useRouter } from "next/router";
-import { MouseEvent, useEffect, useMemo, useRef, useState } from "react";
+import { useRef } from "react";
 
 import { useAuth } from "~community/auth/providers/AuthProvider";
-import AvatarChip from "~community/common/components/molecules/AvatarChip/AvatarChip";
-import AvatarGroup from "~community/common/components/molecules/AvatarGroup/AvatarGroup";
-import Popper from "~community/common/components/molecules/Popper/Popper";
-import SearchBox from "~community/common/components/molecules/SearchBox/SearchBox";
 import AreYouSureModal from "~community/common/components/molecules/AreYouSureModal/AreYouSureModal";
 import Modal from "~community/common/components/organisms/Modal/Modal";
 import { ToastType } from "~community/common/enums/ComponentEnums";
 import ROUTES from "~community/common/constants/routes";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import { useToast } from "~community/common/providers/ToastProvider";
-import { theme } from "~community/common/theme/theme";
 import { AdminTypes } from "~community/common/types/AuthTypes";
-import {
-  AvatarPropTypes,
-  MenuTypes
-} from "~community/common/types/MoleculeTypes";
 import { useGetAttendanceConfiguration } from "~community/attendance/api/AttendanceAdminApi";
-import {
-  useGetSearchedEmployees,
-  useGetEmployeeData
-} from "~community/people/api/PeopleApi";
-import { usePeopleStore } from "~community/people/store/store";
-import {
-  DataFilterEnums,
-  EmploymentStatusTypes
-} from "~community/people/types/EmployeeTypes";
-import { AllEmployeeDataType } from "~community/people/types/PeopleTypes";
 import { useGetWorkLocations, useUpdateWorkLocation } from "~community/configurations/api/WorkLocationApi";
 import GeofenceMap from "~community/configurations/components/molecules/GeofenceMap/GeofenceMap";
+import WorkLocationEmployeeSelector from "~community/configurations/components/molecules/WorkLocationEmployeeSelector/WorkLocationEmployeeSelector";
+import { useUnsavedChangesGuard } from "~community/configurations/hooks/useUnsavedChangesGuard";
 import { WorkLocation, WorkLocationFormValues } from "~community/configurations/types/WorkLocationTypes";
+import { buildWorkLocationValidationSchema } from "~community/common/utils/validationUtils";
 
 interface Props {
   id: number;
@@ -44,13 +26,6 @@ interface Props {
 const UpdateWorkLocation = ({ id }: Props) => {
   const router = useRouter();
   const translateText = useTranslator("configurations", "workLocation");
-  const [isUnsavedModalOpen, setIsUnsavedModalOpen] = useState(false);
-  const [pendingUrl, setPendingUrl] = useState<string | null>(null);
-  const [employeeSearchText, setEmployeeSearchText] = useState("");
-  const [popperOpen, setPopperOpen] = useState(false);
-  const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
-  const boxRef = useRef<HTMLDivElement>(null);
-  const [boxWidth, setBoxWidth] = useState(0);
 
   const { data: locationsPage, isLoading } = useGetWorkLocations("", 0, 100);
   const workLocation: WorkLocation | undefined = locationsPage?.items?.find(
@@ -60,6 +35,7 @@ const UpdateWorkLocation = ({ id }: Props) => {
   const { user } = useAuth();
   const { setToastMessage } = useToast();
   const formikRef = useRef<ReturnType<typeof useFormik<WorkLocationFormValues>> | null>(null);
+  const skipGuardRef = useRef<(() => void) | null>(null);
 
   const { data: attendanceConfig } = useGetAttendanceConfiguration();
 
@@ -68,12 +44,7 @@ const UpdateWorkLocation = ({ id }: Props) => {
       user?.roles?.includes(AdminTypes.ATTENDANCE_ADMIN)) &&
     attendanceConfig?.isGeoFencingEnabled === true;
 
-  const validationSchema = Yup.object({
-    name: Yup.string()
-      .required(translateText(["validation.nameRequired"]))
-      .max(50, translateText(["validation.nameMaxLength"]))
-      .matches(/^[a-zA-Z0-9 ]+$/, translateText(["validation.nameInvalidChars"]))
-  });
+  const validationSchema = buildWorkLocationValidationSchema(translateText);
 
   const { mutate: updateWorkLocation, isPending } = useUpdateWorkLocation(
     () => {
@@ -85,6 +56,7 @@ const UpdateWorkLocation = ({ id }: Props) => {
         isIcon: true
       });
       formikRef.current?.resetForm();
+      skipGuardRef.current?.();
       router.push(`${ROUTES.CONFIGURATIONS.BASE}?tab=organization`);
     },
     () => {
@@ -128,104 +100,9 @@ const UpdateWorkLocation = ({ id }: Props) => {
 
   formikRef.current = formik;
 
-  const { setEmployeeDataParams } = usePeopleStore((state) => state);
-
-  useEffect(() => {
-    setEmployeeDataParams(DataFilterEnums.ACCOUNT_STATUS, [
-      EmploymentStatusTypes.ACTIVE,
-      EmploymentStatusTypes.PENDING
-    ]);
-  }, [setEmployeeDataParams]);
-
-  const { data: employeePages } = useGetEmployeeData();
-  const { data: searchResults } = useGetSearchedEmployees(employeeSearchText);
-
-  const allEmployees: AllEmployeeDataType[] =
-    employeePages?.pages?.flatMap((page: any) => page?.items ?? []) ?? [];
-
-  const displayEmployees = useMemo(() => {
-    return employeeSearchText.length > 0
-      ? ((searchResults ?? []) as AllEmployeeDataType[])
-      : allEmployees;
-  }, [employeeSearchText, searchResults, allEmployees]);
-
-  const selectedIds: number[] = formik.values.employeeIds ?? [];
-  const isAllSelected = formik.values.isAllEmployees;
-
-  const selectedEmployees = useMemo(() => {
-    return selectedIds
-      .map((empId) => allEmployees.find((e) => Number(e.employeeId) === empId))
-      .filter(Boolean) as AllEmployeeDataType[];
-  }, [selectedIds, allEmployees]);
-
-  const selectedCount = isAllSelected
-    ? allEmployees.length
-    : selectedIds.length;
-
-  useEffect(() => {
-    if (boxRef.current) {
-      setBoxWidth(boxRef.current.clientWidth);
-    }
-  }, [popperOpen]);
-
-  const handlePopperClose = () => {
-    setPopperOpen(false);
-    setAnchorEl(null);
-    setEmployeeSearchText("");
-  };
-
-  const handleTriggerClick = (event: MouseEvent<HTMLElement>) => {
-    setEmployeeSearchText("");
-    setAnchorEl(event.currentTarget);
-    setPopperOpen((prev) => !prev);
-  };
-
-  const toggleEmployee = (empId: number) => {
-    if (selectedIds.includes(empId)) {
-      formik.setFieldValue(
-        "employeeIds",
-        selectedIds.filter((id) => id !== empId)
-      );
-    } else {
-      formik.setFieldValue("employeeIds", [...selectedIds, empId]);
-    }
-  };
-
-  const toggleAllEmployees = () => {
-    if (isAllSelected) {
-      formik.setFieldValue("isAllEmployees", false);
-    } else {
-      formik.setFieldValue("isAllEmployees", true);
-      formik.setFieldValue("employeeIds", []);
-    }
-  };
-
-  useEffect(() => {
-    const handleRouteChangeStart = (url: string) => {
-      if (formik.dirty) {
-        setPendingUrl(url);
-        setIsUnsavedModalOpen(true);
-        router.events.emit("routeChangeError");
-        throw "routeChange aborted";
-      }
-    };
-    router.events.on("routeChangeStart", handleRouteChangeStart);
-    return () => {
-      router.events.off("routeChangeStart", handleRouteChangeStart);
-    };
-  }, [formik.dirty, router]);
-
-  const handleResume = () => {
-    setIsUnsavedModalOpen(false);
-    setPendingUrl(null);
-  };
-
-  const handleLeave = () => {
-    setIsUnsavedModalOpen(false);
-    if (pendingUrl) {
-      router.push(pendingUrl);
-    }
-  };
+  const { isUnsavedModalOpen, handleResume, handleLeave, skipNextGuard } =
+    useUnsavedChangesGuard(formik.dirty, router);
+  skipGuardRef.current = skipNextGuard;
 
   if (isLoading) {
     return (
@@ -256,194 +133,7 @@ const UpdateWorkLocation = ({ id }: Props) => {
         />
       </div>
 
-      {/* Employee Selection */}
-      <div>
-        <Typography
-          variant="placeholder"
-          gutterBottom
-          sx={{ color: "common.black" }}
-        >
-          {translateText(["form.assignEmployeesLabel"])}
-        </Typography>
-        <Box
-          ref={boxRef}
-          tabIndex={0}
-          role="combobox"
-          aria-expanded={popperOpen}
-          aria-haspopup="listbox"
-          aria-label={translateText(["form.assignEmployeesLabel"])}
-          sx={{
-            backgroundColor: theme.palette.grey[100],
-            height: "3rem",
-            borderRadius: "0.5rem",
-            display: "flex",
-            alignItems: "center",
-            width: "100%",
-            cursor: "pointer",
-            px: "0.75rem",
-            "&:focus": {
-              outline: `0.125rem solid ${theme.palette.common.black}`,
-              outlineOffset: "-0.125rem"
-            }
-          }}
-          onClick={handleTriggerClick}
-        >
-          {selectedCount === 0 ? (
-            <Typography
-              variant="placeholder"
-              sx={{ color: theme.palette.text.secondary, ml: "0.5rem" }}
-            >
-              {translateText(["form.assignEmployeesLabel"])}
-            </Typography>
-          ) : isAllSelected ? (
-            <AvatarChip
-              firstName="All"
-              lastName="Employees"
-              avatarUrl={undefined}
-              chipStyles={{
-                backgroundColor: "common.white",
-                color: "common.black",
-                height: "2.5rem"
-              }}
-            />
-          ) : selectedCount <= 2 ? (
-            <Box sx={{ display: "flex", gap: "0.5rem" }}>
-              {selectedEmployees.map((emp) => (
-                <AvatarChip
-                  key={emp.employeeId}
-                  firstName={emp.firstName ?? ""}
-                  lastName={selectedCount === 1 ? (emp.lastName ?? "") : ""}
-                  avatarUrl={emp.authPic}
-                  chipStyles={{
-                    backgroundColor: "common.white",
-                    color: "common.black",
-                    height: "2.5rem"
-                  }}
-                />
-              ))}
-            </Box>
-          ) : (
-            <AvatarGroup
-              componentStyles={{
-                ".MuiAvatarGroup-avatar": {
-                  bgcolor: theme.palette.grey[100],
-                  color: theme.palette.primary.dark,
-                  fontSize: "0.875rem",
-                  height: "2.5rem",
-                  width: "2.5rem",
-                  fontWeight: 400
-                }
-              }}
-              avatars={selectedEmployees.map(
-                (emp) =>
-                  ({
-                    firstName: emp.firstName,
-                    lastName: emp.lastName,
-                    image: emp.authPic
-                  }) as AvatarPropTypes
-              )}
-              max={3}
-            />
-          )}
-        </Box>
-
-        <Popper
-          anchorEl={anchorEl}
-          open={popperOpen}
-          position="bottom-end"
-          menuType={MenuTypes.FILTER}
-          id={popperOpen ? "employee-select-popper" : undefined}
-          handleClose={handlePopperClose}
-          timeout={300}
-          containerStyles={{
-            maxHeight: "20.25rem",
-            width: `${boxWidth}px`,
-            backgroundColor: theme.palette.notifyBadge.contrastText,
-            boxShadow: theme.shadows[1],
-            overflow: "hidden"
-          }}
-        >
-          <Box sx={{ p: "0.5rem" }}>
-            <SearchBox
-              placeHolder={translateText(["form.assignEmployeesLabel"])}
-              value={employeeSearchText}
-              setSearchTerm={setEmployeeSearchText}
-              autoFocus
-            />
-          </Box>
-          <Box sx={{ maxHeight: "14rem", overflowY: "auto" }}>
-            {/* All Employees option - always shown first */}
-            <Box
-              sx={{
-                display: "flex",
-                alignItems: "center",
-                px: "0.75rem",
-                py: "0.25rem",
-                cursor: "pointer",
-                "&:hover": { backgroundColor: theme.palette.grey[100] }
-              }}
-              onClick={toggleAllEmployees}
-            >
-              <MuiCheckbox
-                checked={isAllSelected}
-                size="small"
-                sx={{ p: "0.25rem" }}
-              />
-              <AvatarChip
-                firstName="All"
-                lastName="Employees"
-                avatarUrl={undefined}
-                chipStyles={{
-                  backgroundColor: "transparent",
-                  color: "common.black",
-                  height: "2.5rem",
-                  border: "none",
-                  boxShadow: "none"
-                }}
-              />
-            </Box>
-
-            {/* Individual employees */}
-            {!isAllSelected &&
-              displayEmployees.map((emp) => {
-                const empId = Number(emp.employeeId);
-                const isSelected = selectedIds.includes(empId);
-                return (
-                  <Box
-                    key={empId}
-                    sx={{
-                      display: "flex",
-                      alignItems: "center",
-                      px: "0.75rem",
-                      py: "0.25rem",
-                      cursor: "pointer",
-                      "&:hover": { backgroundColor: theme.palette.grey[100] }
-                    }}
-                    onClick={() => toggleEmployee(empId)}
-                  >
-                    <MuiCheckbox
-                      checked={isSelected}
-                      size="small"
-                      sx={{ p: "0.25rem" }}
-                    />
-                    <AvatarChip
-                      firstName={emp.firstName ?? ""}
-                      lastName={emp.lastName ?? ""}
-                      avatarUrl={emp.authPic}
-                      chipStyles={{
-                        backgroundColor: "transparent",
-                        color: "common.black",
-                        height: "2.5rem",
-                        border: "none",
-                        boxShadow: "none"
-                      }}
-                    />
-                  </Box>
-                );
-              })}
-          </Box>
-        </Popper>
-      </div>
+      <WorkLocationEmployeeSelector formik={formik} />
 
       {canSeeGeofence && <GeofenceMap formik={formik} />}
 
@@ -461,6 +151,7 @@ const UpdateWorkLocation = ({ id }: Props) => {
       </div>
 
       <Modal
+        title={translateText(["areYouSureModalTitle"])}
         isModalOpen={isUnsavedModalOpen}
         onCloseModal={handleResume}
         isClosable={false}
